+++ packages/ai/src/models.generated.ts
@@ -13544,13 +13544,13 @@ export const MODELS = {
 			reasoning: true,
 			input: ["text"],
 			cost: {
-				input: 0.84,
-				output: 2.64,
-				cacheRead: 0.156,
+				input: 0.9299999999999999,
+				output: 3,
+				cacheRead: 0.18,
 				cacheWrite: 0,
 			},
 			contextWindow: 1048576,
-			maxTokens: 131072,
+			maxTokens: 32768,
 		} satisfies Model<"openai-completions">,
 		"z-ai/glm-5v-turbo": {
 			id: "z-ai/glm-5v-turbo",